Moved TryCatchNode and CatchNode into Exceptions sub-namespace.
